@@ -1,8 +1,9 @@
-var test = require('tape')
-var Core = require('../../src/core/index.js')
+import test from 'tape'
+import Core from '../../src/core/index.js'
+import russian from '../../src/locales/ru_RU.js'
+import english from '../../src/locales/en_US.js'
 
 test('translation', function (t) {
-  const russian = require('../../src/locales/ru_RU.js')
   const core = new Core({locales: russian})
 
   t.equal(
@@ -15,7 +16,6 @@ test('translation', function (t) {
 })
 
 test('interpolation', function (t) {
-  const english = require('../../src/locales/en_US.js')
   const core = new Core({locales: english})
 
   t.equal(
@@ -28,7 +28,6 @@ test('interpolation', function (t) {
 })
 
 test('pluralization', function (t) {
-  const russian = require('../../src/locales/ru_RU.js')
   const core = new Core({locales: russian})
 
   t.equal(
