@@ -1,6 +1,3 @@
-require('babel/register')({
-  stage: 0
-})
-
+require('babel-register')
 require('./core.spec.js')
 require('./translator.spec.js')
